Removed HOC in favour of inheritance in themed input components. Themed components extracted to separate files
