+++ src/Input/index.js
@@ -1,1 +1,1 @@
-export {default} from './Input.js';
+export {default} from './ThemedInput.js';
